figuring out logic to call get counties
